requirements.txt
gradio==5.41.1
requests==2.32.3
python-dotenv==1.0.1
ollama==0.5.3